Added tools/serve.py
The default HTTP server used by SimpleHTTPServer is single threaded
which can make refreshes of documentation pages a bit slow. This PR
adds a little script which starts a multi-threaded HTTP server for
viewing docs. It also takes the directory to serve on the command
line, and avoids the traceback when Ctrl-C is used.

The "serve" make target has been updated to use the new script.