Cleanup of #1570
According to the python3 docs, it was the python2 range() function that
was deprecated and the python3 range() is equivalent to the python2
xrange function. I've removed the extra characters to reflect this
change.

I feel that python3 being the default for our recommended way to write
charms, this becomes a correct representation of an implementation.